+++ packages/horseman-core/CHANGELOG.md
@@ -1,3 +1,8 @@
+2.8.0 / 2018-1-9
+==================
+- Refactor the ActionFactory to use a fetchResource function. This function has
+  been build to support SSR fetching of resources also.
+
 2.6.1 / 2018-1-8
 ==================
 - Export the thing